+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@ A new header is inserted each time a *tag* is created.
 
 - Fixed morph shapes not rendering in WebGL.
 - Added support for the latest version of emscripten.
+- gltfio: fixed blackness seen with default material.
 
 ## v1.4.2